+++ lib/media_extension_method_channel.dart
@@ -4,7 +4,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 import 'package:media_extension/media_extension_action_types.dart';
 
-import 'media_extension_platform_interface.dart';
+import 'package:media_extension/media_extension_platform_interface.dart';
 
 /// An implementation of [MediaExtensionPlatform] that uses method channels.
 class MethodChannelMediaExtension extends MediaExtensionPlatform {
@@ -12,6 +12,7 @@ class MethodChannelMediaExtension extends MediaExtensionPlatform {
   @visibleForTesting
   final methodChannel = const MethodChannel('media_extension');
 
+  /// This Method retrivies the platform version from native thread.
   @override
   Future<String?> getPlatformVersion() async {
     final version =
@@ -19,9 +20,12 @@ class MethodChannelMediaExtension extends MediaExtensionPlatform {
     return version;
   }
 
+  /// This Method sends the selected Image's [uri] to the native thread.
+  /// from the Flutter thread to send results to apps
+  /// which can handle [ACTION_ATTACH_DATA] Intent of [mimeType] `images/*`.
   @override
   Future<bool> setAs(String uri, String mimeType,
-      {String title = "Set as"}) async {
+      {String title = 'Set as'}) async {
     try {
       final result =
           await methodChannel.invokeMethod('setAs', <String, dynamic>{
@@ -36,6 +40,9 @@ class MethodChannelMediaExtension extends MediaExtensionPlatform {
     return false;
   }
 
+  /// This Method sends the selected Image's [uri] to the native thread.
+  /// from the Flutter thread to send results to apps
+  /// which can handle [ACTION_EDIT] Intent of [mimeType] `images/*`.
   @override
   Future<bool> edit(
     String uri,
@@ -55,6 +62,9 @@ class MethodChannelMediaExtension extends MediaExtensionPlatform {
     return false;
   }
 
+  /// This Method sends the selected Image's [uri] to the native thread.
+  /// from the Flutter thread to send results
+  /// to apps which can handle [ACTION_VIEW] Intent of [mimeType] `images/*`.
   @override
   Future<bool> openWith(
     String uri,
@@ -76,24 +86,29 @@ class MethodChannelMediaExtension extends MediaExtensionPlatform {
   }
 
   /// This Method is triggered by the Native Thread which sends the [intentAction]
-  /// and [uri] information in a HashMap Structure to the Flutter thread.
+  /// and [uri] information in a [HashMap] Structure to the Flutter thread.
   @override
   Future<MediaExtentionAction> getIntentAction() async {
+    //Completer which awaits for the result from native android
+    //thread to send the IntentAction Information.
     final Completer<MediaExtentionAction> completer =
         Completer<MediaExtentionAction>();
     methodChannel.setMethodCallHandler((call) async {
       final dynamic args = call.arguments;
-      final String action = args["action"];
-      final String uri = args["uri"];
+      final String action = args['action'];
+      final String uri = args['uri'];
       MediaExtentionAction intentAction =
-          MediaExtentionAction(action: actionStringify(action), uri: uri);
+          MediaExtentionAction(action: actionParser(action), uri: uri);
+      //Completes the Completer
       completer.complete(intentAction);
     });
+    //Returns the future
     return completer.future;
   }
 
-  /// This Method is sends the selected Image's [uri] to the native thread.
-  /// from the Flutter thread to send results the [INTENT_ACTION_PICK].
+  /// This Method sends the selected Image's [uri] to the native thread.
+  /// from the Flutter thread to send results to the app
+  /// which invoked our app using [ACTION_PICK] INTENT.
   @override
   Future<void> setResult(String uri) async {
     try {